Now all node indexes are actually NodeIndex

+++ zuban_python/src/value/class.rs
@@ -22,7 +22,7 @@ impl<'a> Value<'a> for Class {
     }
 
     fn get_name(&self) -> &'a str {
-        let class_node = unsafe {&*self.file}.tree.get_node_by_index(self.node_index as usize);
+        let class_node = unsafe {&*self.file}.tree.get_node_by_index(self.node_index);
         class_node.get_nth_child(1).get_nth_child(0).get_code()
     }
 }
